# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  CreateUserRequest:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      description:
        example: Senior researcher specializing in quantum computing
        type: string
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        maxLength: 255
        type: string
      password:
        example: PassWord123!
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      research_categories:
        example:
        - QuantumComputing
        - Qiskit
        - Cryogenics
        items:
          type: string
        type: array
      social_media:
        example: '@john on Twitter, linkedin.com/in/john'
        type: string
    required:
    - email
    - name
    - password
    type: object
  ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    required:
    - error
    type: object
...
  UpdateUserRequest:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      description:
        example: Senior researcher specializing in quantum computing
        type: string
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        maxLength: 255
        type: string
      password:
        example: PassWord123!
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      research_categories:
        example:
        - QuantumComputing
        - Qiskit
        - Cryogenics
        items:
          type: string
        type: array
      social_media:
        example: '@john on Twitter, linkedin.com/in/john'
        type: string
    type: object
  UserMembershipResponse:
    properties:
      is_admin:
        example: false
        type: boolean
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      organization_name:
        example: Acme Corp
        type: string
    type: object
  UserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2026-01-01T12:00:00.00000+07:00"
        type: string
      description:
        example: Senior researcher specializing in quantum computing
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: John Doe
        type: string
      organizations:
        items:
          $ref: '#/definitions/UserMembershipResponse'
        type: array
      phone_number:
        example: "+1234567890"
        type: string
      research_categories:
        example:
        - QuantumComputing
        - Qiskit
        - Cryogenics
        items:
          type: string
        type: array
      social_media:
        example: '@john on Twitter, linkedin.com/in/john'
        type: string
      updated_at:
        example: "2026-01-01T12:00:00.00000+07:00"
        type: string
    required:
    - created_at
    - email
    - id
    - name
    - research_categories
    - updated_at
    type: object
info:
  contact: {}
  description: A simple REST API for managing users and organizations
  title: Customer API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with email, name, organization, and optional
        details
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a user by ID
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get a user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update an existing user by ID (partial updates supported)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update a user
      tags:
      - users
  /users/organization/{org_id}:
    get:
      consumes:
      - application/json
      description: Get all users in a specific organization
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get users by organization
      tags:
      - users
  /users/search:
    get:
      consumes:
      - application/json
      description: Search users by name or email using ILIKE query
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: 'Maximum number of results to return (default: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Search users
      tags:
      - users
swagger: "2.0"
...